Moved Algae and Vision into their own folders in the subsystem folder.
