+++ src/main/java/frc/robot/subsystems/vision/LimelightHelpers.java
@@ -1,6 +1,6 @@
 // LimelightHelpers v1.10 (REQUIRES LLOS 2024.9.1 OR LATER)
 
-package frc.robot;
+package frc.robot.subsystems.vision;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonFormat.Shape;
