+++ src/main/java/frc/robot/subsystems/vision/VisionSubsystem.java
@@ -1,4 +1,4 @@
-package frc.robot;
+package frc.robot.subsystems.vision;
 
 import edu.wpi.first.wpilibj2.command.*;
